+++ src/libs/hooks/cart/queries.ts
@@ -1,5 +1,5 @@
-import { getCartItemList } from '@/libs/api/cart/api';
-import { CartItem } from '@/types/cartItems';
+import { getCartItemList } from '@/libs/api/cart/cart-api';
+import { CartItem } from '@/types/cart-items';
 import { useQuery } from '@tanstack/react-query';
 
 export const useGetCartItems = (userId: CartItem['userId']) => {
